remove uppercase
